// src/sdk/requests/services.gen.ts
export class PatientControllerService {
  /**
   * @param data The data for the request.
   * @param data.id
   * @returns PatientDTO OK
   * @throws ApiError
   */
  public static getPatientById(
    data: GetPatientByIdData,
  ): CancelablePromise<GetPatientByIdResponse> {
    return __request(OpenAPI, {
      method: "GET",
      url: "/api/v1/patient/{id}",
      path: {
        id: data.id,
      },
    });
  }

  /**
   * @param data The data for the request.
   * @param data.id
   * @param data.requestBody
   * @returns string OK
   * @throws ApiError
   */
  public static changePatient(
    data: ChangePatientData,
  ): CancelablePromise<ChangePatientResponse> {
    return __request(OpenAPI, {
      method: "PUT",
      url: "/api/v1/patient/{id}",
      path: {
        id: data.id,
      },
      body: data.requestBody,
      mediaType: "application/json",
    });
  }

  /**
   * @param data The data for the request.
   * @param data.id
   * @returns unknown OK
   * @throws ApiError
   */
  public static deletePatient(
    data: DeletePatientData,
  ): CancelablePromise<DeletePatientResponse> {
    return __request(OpenAPI, {
      method: "DELETE",
      url: "/api/v1/patient/{id}",
      path: {
        id: data.id,
      },
    });
  }

  /**
   * @param data The data for the request.
   * @param data.requestBody
   * @returns PatientDTO OK
   * @throws ApiError
   */
  public static registerNewPatient(
    data: RegisterNewPatientData,
  ): CancelablePromise<RegisterNewPatientResponse> {
    return __request(OpenAPI, {
      method: "POST",
      url: "/api/v1/patient/register",
      body: data.requestBody,
      mediaType: "application/json",
    });
  }

  /**
   * @returns PatientDTO OK
   * @throws ApiError
   */
  public static getMyProfile(): CancelablePromise<GetMyProfileResponse> {
    return __request(OpenAPI, {
      method: "GET",
      url: "/api/v1/patient/me",
    });
  }

  /**
   * @param data The data for the request.
   * @param data.requestBody
   * @returns PatientDTO OK
   * @throws ApiError
   */
  public static updatePatient(
    data: UpdatePatientData,
  ): CancelablePromise<UpdatePatientResponse> {
    return __request(OpenAPI, {
      method: "PATCH",
      url: "/api/v1/patient/me",
      body: data.requestBody,
      mediaType: "application/json",
    });
  }

  /**
   * @returns Response OK
   * @throws ApiError
   */
  public static get(): CancelablePromise<GetResponse> {
    return __request(OpenAPI, {
      method: "GET",
      url: "/api/v1/patient/patients",
    });
  }

  /**
   * @returns PatientDashboardDTO OK
   * @throws ApiError
   */
  public static getDashboard(): CancelablePromise<GetDashboardResponse> {
    return __request(OpenAPI, {
      method: "GET",
      url: "/api/v1/patient/dashboard",
    });
  }
}

export class AdminControllerService {
  /**
   * @returns Response OK
   * @throws ApiError
   */
  public static getAdminProfile(): CancelablePromise<GetAdminProfileResponse> {
    return __request(OpenAPI, {
      method: "GET",
      url: "/api/v1/admin/profile",
    });
  }

  /**
   * @param data The data for the request.
   * @param data.requestBody
   * @returns Response OK
   * @throws ApiError
   */
  public static updateAdminProfile(
    data: UpdateAdminProfileData,
  ): CancelablePromise<UpdateAdminProfileResponse> {
    return __request(OpenAPI, {
      method: "PUT",
      url: "/api/v1/admin/profile",
      body: data.requestBody,
      mediaType: "application/json",
    });
  }
}
// ...
export class PrescriptionControllerService {
  /**
   * @param data The data for the request.
   * @param data.requestBody
   * @returns Response OK
   * @throws ApiError
   */
  public static createPrescription(
    data: CreatePrescriptionData,
  ): CancelablePromise<CreatePrescriptionResponse> {
    return __request(OpenAPI, {
      method: "POST",
      url: "/api/v1/prescription",
      body: data.requestBody,
      mediaType: "application/json",
    });
  }

  /**
   * @param data The data for the request.
   * @param data.uuid
   * @returns Response OK
   * @throws ApiError
   */
  public static getPrescriptionByUuid(
    data: GetPrescriptionByUuidData,
  ): CancelablePromise<GetPrescriptionByUuidResponse> {
    return __request(OpenAPI, {
      method: "GET",
      url: "/api/v1/prescription/{uuid}",
      path: {
        uuid: data.uuid,
      },
    });
  }

  /**
   * @param data The data for the request.
   * @param data.appointmentUuid
   * @returns Response OK
   * @throws ApiError
   */
  public static getPrescriptionByAppointmentUuid(
    data: GetPrescriptionByAppointmentUuidData,
  ): CancelablePromise<GetPrescriptionByAppointmentUuidResponse> {
    return __request(OpenAPI, {
      method: "GET",
      url: "/api/v1/prescription/appointment/{appointmentUuid}",
      path: {
        appointmentUuid: data.appointmentUuid,
      },
    });
  }
}
// ...
export class AppointmentControllerService {
  /**
   * @param data The data for the request.
   * @param data.requestBody
   * @returns Response OK
   * @throws ApiError
   */
  public static createNewAppointment(
    data: CreateNewAppointmentData,
  ): CancelablePromise<CreateNewAppointmentResponse> {
    return __request(OpenAPI, {
      method: "POST",
      url: "/api/v1/appointment",
      body: data.requestBody,
      mediaType: "application/json",
    });
  }

  /**
   * @param data The data for the request.
   * @param data.page Zero-based page index (0..N)
   * @param data.size The size of the page to be returned
   * @param data.sort Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.
   * @returns Response OK
   * @throws ApiError
   */
  public static getPatientAppointments(
    data: GetPatientAppointmentsData = {},
  ): CancelablePromise<GetPatientAppointmentsResponse> {
    return __request(OpenAPI, {
      method: "GET",
      url: "/api/v1/appointment/patient",
      query: {
        page: data.page,
        size: data.size,
        sort: data.sort,
      },
    });
  }

  /**
   * @param data The data for the request.
   * @param data.page Zero-based page index (0..N)
   * @param data.size The size of the page to be returned
   * @param data.sort Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.
   * @returns Response OK
   * @throws ApiError
   */
  public static getDoctorAppointments(
    data: GetDoctorAppointmentsData = {},
  ): CancelablePromise<GetDoctorAppointmentsResponse> {
    return __request(OpenAPI, {
      method: "GET",
      url: "/api/v1/appointment/doctor",
      query: {
        page: data.page,
        size: data.size,
        sort: data.sort,
      },
    });
  }

  /**
   * @param data The data for the request.
   * @param data.page Zero-based page index (0..N)
   * @param data.size The size of the page to be returned
   * @param data.sort Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.
   * @returns Response OK
   * @throws ApiError
   */
  public static getAllAppointments(
    data: GetAllAppointmentsData = {},
  ): CancelablePromise<GetAllAppointmentsResponse> {
    return __request(OpenAPI, {
      method: "GET",
      url: "/api/v1/appointment/admin",
      query: {
        page: data.page,
        size: data.size,
        sort: data.sort,
      },
    });
  }
}

export class DoctorControllerService {
  /**
   * @returns Response OK
   * @throws ApiError
   */
  public static getDoctorProfile(): CancelablePromise<GetDoctorProfileResponse> {
    return __request(OpenAPI, {
      method: "GET",
      url: "/api/v1/doctor/profile",
    });
  }

  /**
   * @param data The data for the request.
   * @param data.requestBody
   * @returns Response OK
   * @throws ApiError
   */
  public static updateDoctorProfile(
    data: UpdateDoctorProfileData,
  ): CancelablePromise<UpdateDoctorProfileResponse> {
    return __request(OpenAPI, {
      method: "PATCH",
      url: "/api/v1/doctor/profile",
      body: data.requestBody,
      mediaType: "application/json",
    });
  }

  /**
   * @returns Response OK
   * @throws ApiError
   */
  public static getAllDoctors(): CancelablePromise<GetAllDoctorsResponse> {
    return __request(OpenAPI, {
      method: "GET",
      url: "/api/v1/doctor",
    });
  }

  /**
   * @returns Response OK
   * @throws ApiError
   */
  public static getDoctorDashboard(): CancelablePromise<GetDoctorDashboardResponse> {
    return __request(OpenAPI, {
      method: "GET",
      url: "/api/v1/doctor/dashboard",
    });
  }
}